lint: Fix linting errors 	- Updated eslint config to ignore trailing commas and fix no-shadow TypeScript issue 	- Changed object index signatures to use `Record` 	- Removed string unions and decided to use explicit defined strings ("US") instead 	- Changed #snowflakeToDate parameter to be of type `bigint` instead of `BigInt`

+++ src/utils/DiscordUtils.ts
@@ -1,11 +1,11 @@
 // https://discord.com/developers/docs/reference#snowflakes
-const snowflakeToDate = (snowflake: string | BigInt): Date => {
+const snowflakeToDate = (snowflake: string | bigint): Date => {
 	if (typeof snowflake === "string") {
 		snowflake = BigInt(snowflake);
 	}
 	// (snowflake >> 22) + 1420070400000
 	// Make sure to cast 22 to BigInt too by doing 22n
-	return new Date(Number((snowflake as bigint) >> 22n) + 1420070400000);
+	return new Date(Number(snowflake >> 22n) + 1420070400000);
 };
 
 export const DiscordUtils = {
